requirements.txt
# Core app
streamlit==1.37.1
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ML + DSP
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
librosa==0.10.2.post1
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
matplotlib==3.9.0
moviepy==1.0.3

# Utils
python-multipart==0.0.9
pydantic==2.8.2
